bharatfd/requirements.txt
anyio
asgiref
async-timeout
certifi
Django
django-ckeditor
django-js-asset
django-redis
djangorestframework
exceptiongroup
googletrans
gunicorn
h11
h2
hpack
httpcore
httpx
hyperframe
idna
packaging
python-dotenv
redis
sniffio
sqlparse
typing_extensions
pytest
pytest-django